+++ SharpQMapParser/Core/Point.cs
@@ -2,9 +2,9 @@
 {
     public struct Point
     {
-        public int x, y, z;
+        public float x, y, z;
 
-        public Point(int x, int y, int z)
+        public Point(float x, float y, float z)
         {
             this.x = x;
             this.y = y;
